Move current context to a separate module.
